+++ Example/Example/ifaddrs.swift
@@ -14,10 +14,10 @@ import Foundation
 enum Ifaddrs_h {
     /**
      using `getifaddrs`, find the `en0` interface and return its IP address.
-    
+
      Adapted from code on StackOverflow, primarily:
      https://stackoverflow.com/a/30754194/1196205
-    
+
      - returns: Device's WIFI IP address.
      */
     static func getWifiIPAddress() -> String? {