+++ Package.swift
@@ -15,8 +15,8 @@ let package = Package(
   targets: [
     .binaryTarget(
       name: "StripeTerminal",
-      url: "https://github.com/stripe/stripe-terminal-ios/releases/download/5.7.0/StripeTerminal.xcframework.zip",
-      checksum: "2335ca022c842bf73a6e098df653934805df36b82335fcb7a6a5798b11c79ffb"
+      url: "https://github.com/stripe/stripe-terminal-ios/releases/download/5.8.0/StripeTerminal.xcframework.zip",
+      checksum: "92c65013c4d92ce9dcba2310a0103651756bb65eeca409324ec54ed9896da4f7"
     )
   ]
 )